+++ Runtime/SubAssetContainer.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace nadena.dev.ndmf.runtime
+{
+    /// <summary>
+    /// This ScriptableObject is used as the root asset when storing generated assets.
+    /// </summary>
+    [PreferBinarySerialization]
+    public class SubAssetContainer : ScriptableObject
+    {
+    }
+}